Rename data_structure module to data_structure_field for clarity
